FIX: Mudando a estrutura de pastas de board e criando o botão para adicionar um novo board
