--- frontend/src/app/features/board/board.component.html
@@ -1,1 +1,0 @@
-<p>board.component works!</p>
