--- frontend/src/app/features/board/board.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-board.component',
-  imports: [],
-  templateUrl: './board.component.html',
-  styleUrl: './board.component.scss'
-})
-export class BoardComponent {
-
-}
