+++ frontend/src/app/features/board/list/board.component.html
@@ -1,0 +1,1 @@
+<button (click)="putBoard($event)">Adicionar um Quadro</button>
