+++ frontend/src/app/shared/DTO/board.dto.ts
@@ -1,0 +1,7 @@
+export interface IBoard {
+    id: number;
+    title: string;
+    userId: number;
+    createdAt: Date;
+    updatedAt: Date;
+}
